requirements.txt
# Core dependencies for EPANET Leak Detection
# Python 3.9+

# Data processing
pandas>=1.3.0
numpy>=1.21.0
pyarrow>=6.0.0  # for Parquet support

# EPANET simulation
wntr>=0.4.0

# Machine Learning
scikit-learn>=1.0.0
catboost>=1.1.0  # GPU support included
lightgbm>=3.3.0
imbalanced-learn>=0.9.0  # for SMOTE

# Hyperparameter optimization (optional)
optuna>=3.0.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Utilities
pyyaml>=6.0
tqdm>=4.62.0

# API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
requests>=2.31.0

# Jupyter (for notebook execution)
jupyter>=1.0.0
nbconvert>=6.5.0
ipykernel>=6.0.0
